backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
ccxt==4.3.45
httpx==0.27.0
supabase==2.4.6
apscheduler==3.10.4
python-telegram-bot==21.3
pydantic==2.7.1
pydantic-settings==2.3.0
cachetools==5.3.3
asyncio-throttle==1.0.2
websockets==12.0
python-jose==3.3.0
